+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Added
+
+- Add support for device detection (`device`)
+
 ## 1.1.0 (2023-10-31)
 
 ### Added